extractChoose #1 - New package games.strategy.triplea.ui.chooser - Move ResourceChooser and PlayerChooser to new package
