+++ game-app/game-headed/src/main/java/games/strategy/triplea/ui/EditPanel.java
@@ -25,6 +25,8 @@ import games.strategy.triplea.delegate.move.validation.MoveValidator;
 import games.strategy.triplea.delegate.power.calculator.CombatValueBuilder;
 import games.strategy.triplea.delegate.remote.IEditDelegate;
 import games.strategy.triplea.formatter.MyFormatter;
+import games.strategy.triplea.ui.chooser.PlayerChooser;
+import games.strategy.triplea.ui.chooser.ResourceChooser;
 import games.strategy.triplea.ui.panels.map.MapSelectionListener;
 import games.strategy.triplea.ui.panels.map.UnitSelectionListener;
 import games.strategy.triplea.util.TransportUtils;
